Replaced the crc function. Now can communicate with the PC.

+++ platformio/src/packet_crc.h
@@ -4,6 +4,6 @@
 
 namespace packet_crc {
   
-uint16_t gen_crc16(const uint8_t *data, uint16_t size);
+uint16_t gen_crc16(const uint8_t *data, int size);
 
 } // namespace packet_crc
